refactor: fetch menuitems query refactored
